refactor FormObj, IssueAnnotation
