+++ Reported/ViewControllers/FormViewController.swift
@@ -12,35 +12,6 @@ import MapKit
 import Parse
 import DropDown
 
-final class FormObj: Codable {
-    var issueCategory: String
-    var dirOfTravel: String
-    var transMode: String
-    var nearestCrossStreet: String
-    var dateTime: String
-    var latitude: String
-    var longitude: String
-    var descripText: String
-    var name: String
-    var email: String
-    var phone: String
-    var followUp: Bool
-    
-    init(issueCategory: String!, dirOfTravel: String!, transMode: String!, nearestCrossStreet: String!, dateTime: String!, latitude: String!, longitude: String!, descripText: String!, name: String!, email: String!, phone: String!, followUp: Bool!) {
-        self.issueCategory = issueCategory
-        self.dirOfTravel = dirOfTravel
-        self.transMode = transMode
-        self.nearestCrossStreet = nearestCrossStreet
-        self.dateTime = dateTime
-        self.latitude = latitude
-        self.longitude = longitude
-        self.descripText = descripText
-        self.name = name
-        self.email = email
-        self.phone = phone
-        self.followUp = followUp
-    }
-}
 
 struct Form: Codable {
     var issueCategory: String
@@ -217,7 +188,7 @@ class FormViewController: UIViewController,UIImagePickerControllerDelegate, UINa
             
             // Form Codable
             let form = Form(issueCategory: issueTextField.text!, dirOfTravel: dirTextField.text!, transMode: transTextField.text!, nearestCrossStreet: streetTextField.text!, dateTime: datePickerTextField.text!, latitude: pinLocation!.latitude, longitude: pinLocation!.longitude, descripText: descipTextField.text!, name: "test", email: "email", phone: "phone", followUp: followUpSwitch.isOn)
-            
+        
             // Send to Parse
             postParse(form: form)
